client/filestructure: move confirmation modal
